jobs/requirements.txt
prefect==1.4.1
requests
pandas
google-cloud-storage
google-cloud-bigquery
dbt-core>=1.3.0
dbt-bigquery>=1.3.0
